move: divide file structure
core/ contains only essential classes.
data-structure/ contains only data classes such as `Vector`, `Matrix`, etc.
fx/ contains only effect classes.
