+++ src/engine/data-structure/vector.js
